requirements.txt
asgiref==3.4.1
cffi==1.15.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.12
django-heroku==0.3.1
django-login-history==0.1.1
django-oauth-toolkit==1.7.0
django-password-validators==1.6.1
django-user-sessions==1.7.1
djangorestframework==3.13.1
drf-history==1.0.1
drf-yasg==1.20.0
gunicorn==20.1.0
inflection==0.5.1
itypes==1.2.0
jwcrypto==1.0
oauthlib==3.2.0
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
social-auth-app-django==5.0.0
social-auth-core==4.2.0
typing-extensions==4.1.1
uritemplate==4.1.1
whitenoise==5.3.0
